+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- `taskCompletionTimestamp` to `OneBlinkUploader.uploadSubmission()` options
+
+### Added
+
 - `pdfConversionId` and `pdfConversionJobId` to `OneBlinkUploader.uploadPDFConversion()` result
 - `OneBlinkUploader.uploadFormCustomPDF()`