+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `username` prop to `uploadPrefillData`
+
 ## [6.0.0] - 2026-03-31
 
 ### Added