+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-## [6.0.0] - 2026-03-31
-
 ### Added
 
 - `taskCompletionTimestamp` to `OneBlinkUploader.uploadSubmission()` options
+
+## [6.0.0] - 2026-03-31
 
 ### Added